% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{aggregate_counts}
\alias{aggregate_counts}
\title{Aggregates reads counts from SNPs to transcript and calculates ratios}
\usage{
aggregate_counts(df, threshold = 10)
}
\arguments{
\item{df}{dataframe with count in columns 2 and 3}
}
\value{
dataframe
}
\description{
Aggregates reads counts from SNPs to transcript and calculates ratios
}
