% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{analyze_imbalance_vs_counts}
\alias{analyze_imbalance_vs_counts}
\title{Reads ratios from the file and calculate gini indexes}
\usage{
analyze_imbalance_vs_counts(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
ggplot2
}
\description{
Reads ratios from the file and calculate gini indexes
}
