% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_NPCs.R
\name{analyze_NPC_pulled_clones}
\alias{analyze_NPC_pulled_clones}
\title{Runs analysis for NPCs: pull clones, look at SDs}
\usage{
analyze_NPC_pulled_clones()
}
\value{
ggplot
}
\description{
Runs analysis for NPCs: pull clones, look at SDs
}
