% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{analyze_rin}
\alias{analyze_rin}
\title{Analyze rin vs. sd for all individuals for a given tissue}
\usage{
analyze_rin(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
ggplot2
}
\description{
Analyze rin vs. sd for all individuals for a given tissue
}
