% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{analyze_sd_by_threshold}
\alias{analyze_sd_by_threshold}
\title{Reads counts, apply thresholds, plot distributions}
\usage{
analyze_sd_by_threshold(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
ggplot2
}
\description{
Reads counts, apply thresholds, plot distributions
}
