% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{analyze_sd_for_tissues}
\alias{analyze_sd_for_tissues}
\title{Analyze sd distributions for all individuals for all tissues}
\usage{
analyze_sd_for_tissues()
}
\value{
ggplot2
}
\description{
Analyze sd distributions for all individuals for all tissues
}
