% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{analyze_sd_MAE_BAE}
\alias{analyze_sd_MAE_BAE}
\title{Reads ratios from the file and analyze standard deviations}
\usage{
analyze_sd_MAE_BAE(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
ggplot2
}
\description{
Reads ratios from the file and analyze standard deviations
}
