% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_tissueSiteDeatilIds.R
\name{available_tissueSiteDetailIds}
\alias{available_tissueSiteDetailIds}
\title{Available tissue types}
\usage{
available_tissueSiteDetailIds()
}
\value{
A tibble.
}
\description{
Tissue IDs that can be supplied to functions which include the argument
\code{tissueSiteDetailId}.
}
\examples{
available_tissueSiteDetailIds()
}
\concept{Utilities}
