% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_eqtl.R
\name{calculate_eqtl}
\alias{calculate_eqtl}
\title{Calculate Expression Quantitative Trait Loci}
\usage{
calculate_eqtl(tissueSiteDetailId, gencodeId, variantId, datasetId = "gtex_v8")
}
\arguments{
\item{tissueSiteDetailId}{String. The ID of the tissue of interest. Can be a
GTEx specific ID or an Ontology ID.}

\item{gencodeId}{String. A Versioned GENCODE ID of a gene, e.g.
ENSG00000065613.9.}

\item{variantId}{String. A gtex variant ID.}

\item{datasetId}{String (default = "gtex_v8"). Options: "gtex_v8" or
"gtex_snrnaseq_pilot". Unique identifier of a dataset. Usually includes a
data source and data release.}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Dynamic-Association-Endpoints/operation/calculate_expression_quantitative_trait_loci_api_v2_association_dyneqtl_get}{Calculate your own eQTLs}
}
\examples{
# perform request
result <- calculate_eqtl(tissueSiteDetailId = "Whole_Blood",
               gencodeId = "ENSG00000203782.5",
               variantId = "rs79641866")

# returns a tibble with a single row
result

# unnest list columns with tidyr::unnest()
tidyr::unnest(result, c("data", "genotypes"))
}
\concept{Dynamic Association Endpoints}
