% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_eqtls.R
\name{calculate_eqtls}
\alias{calculate_eqtls}
\title{Calculate Expression Quantitative Trait Loci}
\usage{
calculate_eqtls(
  tissueSiteDetailId,
  gencodeId,
  variantId,
  datasetId = "gtex_v8"
)
}
\arguments{
\item{tissueSiteDetailId}{String. The ID of the tissue of interest. Can be a
GTEx specific ID (e.g. "Whole_Blood"; use \code{\link[=available_tissueSiteDetailIds]{available_tissueSiteDetailIds()}}
to see valid values) or an Ontology ID.}

\item{gencodeId}{String. A Versioned GENCODE ID of a gene, e.g.
"ENSG00000065613.9".}

\item{variantId}{String. A gtex variant ID.}

\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".
Default = "gtex_v8".}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Dynamic-Association-Endpoints/operation/calculate_expression_quantitative_trait_loci_api_v2_association_dyneqtl_get}{Calculate your own eQTLs}
}
\details{
Notes on output:
\itemize{
\item Beta and standard error are recorded in columns \code{nes} and \code{error} respectively (see \href{https://gtexportal.org/home/faq#nes_beta}{GTEx FAQs})
\item \code{variantId} contains (in order) chromosome, position, reference allele, alternative allele and human genome build separated by underscores. The reference and alternative alleles for "chr1_13550_G_A_b38" for example are "G" and "A" respectively.
}

Notes on input:
\itemize{
\item Argument \code{variantId} also accepts RSIDs.
}
}
\examples{
# perform request - returns a tibble with a single row
calculate_eqtls(tissueSiteDetailId = "Whole_Blood",
                gencodeId = "ENSG00000203782.5",
                variantId = "rs79641866")

# unnest list columns with tidyr::unnest()
calculate_eqtls(tissueSiteDetailId = "Whole_Blood",
                gencodeId = "ENSG00000203782.5",
                variantId = "rs79641866") |>
  tidyr::unnest(c("data", "genotypes"))
}
\seealso{
Other Dynamic Association Endpoints: 
\code{\link{calculate_ieqtls}()}
}
\concept{Dynamic Association Endpoints}
