% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{compare_high_vs_low_samples}
\alias{compare_high_vs_low_samples}
\title{Compares sample with high and low ratio variability}
\usage{
compare_high_vs_low_samples(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
ggplot2
}
\description{
Compares sample with high and low ratio variability
}
