% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{compare_same_ind_diff_tissues}
\alias{compare_same_ind_diff_tissues}
\title{Compare same individuals in different tissues}
\usage{
compare_same_ind_diff_tissues(sample_ids)
}
\arguments{
\item{sample_ids}{ids for samples of interest}
}
\value{
ggplot2
}
\description{
Compare same individuals in different tissues
}
