% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runForTissue.R
\name{generatePlots}
\alias{generatePlots}
\title{Generate variatility plots for a given tissue}
\usage{
generatePlots(name, mode)
}
\arguments{
\item{name}{tissue name}

\item{mode}{"norm_counts" / "log" meaning using normalized by library size counts or using log(normalized_counts)}
}
\value{
None
}
\description{
Generate variatility plots for a given tissue
}
