% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{get_ASE_files_list}
\alias{get_ASE_files_list}
\title{Gets a list of files containing ASE information for a given tissue}
\usage{
get_ASE_files_list(fullName)
}
\arguments{
\item{fullName}{Tissue name}
}
\value{
dataframe
}
\description{
Gets a list of files containing ASE information for a given tissue
}
