% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_classification}
\alias{get_classification}
\title{Returns tissue classification with Ensembl id column}
\usage{
get_classification(name,
  path_to_class = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/classification/")
}
\arguments{
\item{name}{Short tissue name}

\item{path_to_class}{Path to file name_class.txt with MAE/BAE classification call (output by MAGIC or RNA-seq analysis)}
}
\value{
tissue code
}
\description{
Returns tissue classification with Ensembl id column
}
