% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clustered_median_exon_expression.R
\name{get_clustered_median_exon_expression}
\alias{get_clustered_median_exon_expression}
\title{Get Clustered Median Exon Expression}
\usage{
get_clustered_median_exon_expression(
  tissueSiteDetailIds,
  datasetId = "gtex_v8",
  sampleId = "^GTEX-[A-Z0-9]{5}-[0-9]{4}-SM-[A-Z0-9]{5}$",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{tissueSiteDetailIds}{Character vector of IDs for tissues of interest.
Can be GTEx specific IDs (e.g. "Whole_Blood"; use
\code{\link[=available_tissueSiteDetailIds]{available_tissueSiteDetailIds()}} to see valid values) or Ontology IDs.}

\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".
Default = "gtex_v8".}

\item{sampleId}{String. \verb{^GTEX-[A-Z0-9]\{5\}-[0-9]\{4\}-SM-[A-Z0-9]\{5\}$}}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A Tibble
}
\description{
\itemize{
\item Find median transcript expression data along with hierarchical clusters .
\item Returns median normalized transcript expression in tissues of all known transcripts of a given gene along with the hierarchical clustering results of tissues and transcripts, based on exon expression, in Newick format.
\item The hierarchical clustering is performed by calculating Euclidean distances and using the average linkage method.
\item This endpoint is not paginated.
\item By default, this endpoint queries the latest GTEx release.
}

\href{https://gtexportal.org/api/v2/redoc#tag/Expression-Data-Endpoints/operation/get_clustered_median_exon_expression_api_v2_expression_clusteredMedianExonExpression_get}{GTEx Portal API documentation}
}
\examples{
\dontrun{
get_clustered_median_exon_expression(tissueSiteDetailIds = "Artery_Aorta")
}
}
\seealso{
Other Expression Data Endpoints: 
\code{\link{get_clustered_median_gene_expression}()},
\code{\link{get_clustered_median_junction_expression}()},
\code{\link{get_clustered_median_transcript_expression}()},
\code{\link{get_expression_pca}()},
\code{\link{get_gene_expression}()},
\code{\link{get_median_exon_expression}()},
\code{\link{get_median_gene_expression}()},
\code{\link{get_median_junction_expression}()},
\code{\link{get_median_transcript_expression}()},
\code{\link{get_single_nucleus_gex_summary}()},
\code{\link{get_single_nucleus_gex}()},
\code{\link{get_top_expressed_genes}()}
}
\concept{Expression Data Endpoints}
