% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collapsed_gene_model_exon.R
\name{get_collapsed_gene_model_exon}
\alias{get_collapsed_gene_model_exon}
\title{Get Collapsed Gene Model Exon}
\usage{
get_collapsed_gene_model_exon(
  gencodeIds,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{gencodeIds}{A character vector of Versioned GENCODE IDs, e.g.
c("ENSG00000132693.12", "ENSG00000203782.5").}

\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".
Default = "gtex_v8".}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A Tibble
}
\description{
\itemize{
\item This service returns the collapsed exons in the gene model of the given gene.
\item Gene-level and exon-level expression quantification were based on the GENCODE
annotation, collapsed to a single transcript model for each gene using an
algorithm developed by the GTEx analysis team.
\item By default, this service queries the models used by the latest GTEx release.
}

\href{https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_collapsed_gene_model_exon_api_v2_dataset_collapsedGeneModelExon_get}{GTEx Portal API documentation}
}
\examples{
get_collapsed_gene_model_exon(gencodeIds = c("ENSG00000132693.12",
                                                  "ENSG00000203782.5"))
}
\seealso{
Other Datasets Endpoints: 
\code{\link{get_annotation}()},
\code{\link{get_downloads_page_data}()},
\code{\link{get_file_list}()},
\code{\link{get_full_get_collapsed_gene_model_exon}()},
\code{\link{get_functional_annotation}()},
\code{\link{get_linkage_disequilibrium_by_variant_data}()},
\code{\link{get_linkage_disequilibrium_data}()},
\code{\link{get_sample}()},
\code{\link{get_subject}()},
\code{\link{get_tissue_site_detail}()},
\code{\link{get_variant_by_location}()},
\code{\link{get_variant}()}
}
\concept{Datasets Endpoints}
