% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{get_counts_for_sample}
\alias{get_counts_for_sample}
\title{Gets a table with reference and allelic counts for a given sample and given tissue}
\usage{
get_counts_for_sample(sample_id, tissue)
}
\arguments{
\item{sample_id}{sample ID of interest}

\item{tissue}{tissue of interest}
}
\value{
dataframe
}
\description{
Gets a table with reference and allelic counts for a given sample and given tissue
}
