% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processToCounts.R
\name{get_counts_statistics}
\alias{get_counts_statistics}
\title{Calculates mean, sd, CV, CV2 for each gene}
\usage{
get_counts_statistics(sv, min_counts_per_gene = 10)
}
\arguments{
\item{sv}{ExpressionSet for a given tissue}

\item{min_counts_per_gene}{Minimum number of counts per gene to be considered}
}
\value{
dataframe with counts statistics
}
\description{
Calculates mean, sd, CV, CV2 for each gene
}
