% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processToCounts.R
\name{get_full_name}
\alias{get_full_name}
\title{Gets full tissue name from a short name}
\usage{
get_full_name(name)
}
\arguments{
\item{name}{Short tissue name}
}
\value{
full tissue name
}
\description{
Gets full tissue name from a short name
}
