% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_search.R
\name{get_gene_search}
\alias{get_gene_search}
\title{Get Gene Search}
\usage{
get_gene_search(
  geneId,
  gencodeVersion = "v26",
  genomeBuild = "GRCh38/hg38",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{geneId}{String.}

\item{gencodeVersion}{String (default = "v26"). GENCODE annotation release.
Either "v26" or "v19".}

\item{genomeBuild}{String (default = "GRCh38/hg38"). Either "GRCh38/hg38" or
"GRCh37/hg19".}

\item{page}{Integer.}

\item{itemsPerPage}{Integer.}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Reference-Genome-Endpoints/operation/get_gene_search_api_v2_reference_geneSearch_get}{Find genes that are partial or complete match of a gene_id}
}
\details{
\itemize{
\item gene_id could be a gene symbol, a gencode ID, or an Ensemble ID
\item Gencode Version and Genome Build must be specified
}
}
\examples{
get_gene_search("CRP")
}
\concept{Reference Genome Endpoints}
