% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{\href{https://gtexportal.org/api/v2/redoc#tag/Reference-Genome-Endpoints/operation/get_genes_api_v2_reference_gene_get}{Get Genes}}
\usage{
get_genes(
  geneId,
  gencodeVersion = "v26",
  genomeBuild = "GRCh38/hg38",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{geneId}{String.}

\item{gencodeVersion}{String (default = "v26"). GENCODE annotation release.
Either "v26" or "v19".}

\item{genomeBuild}{String (default = "GRCh38/hg38"). Either "GRCh38/hg38" or
"GRCh37/hg19".}

\item{page}{Integer.}

\item{itemsPerPage}{Integer.}
}
\value{
A tibble
}
\description{
This service returns information about reference genes. A genome build and
GENCODE version must be provided.
}
\details{
\itemize{
\item Genes are searchable by gene symbol, GENCODE ID and versioned GENCODE ID.
\item Versioned GENCODE ID is recommended to ensure unique ID matching.
\item By default, this service queries the genome build and GENCODE version used by the latest GTEx release.
}
}
\examples{
get_genes("CRP")
}
\seealso{
Other Reference Genome Endpoints: 
\code{\link{get_gene_search}()}
}
\concept{Reference Genome Endpoints}
