% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Get Image}
\usage{
get_image(tissueSampleId = NULL, page = 0, itemsPerPage = 250)
}
\arguments{
\item{tissueSampleId}{Array of strings. A list of Tissue Sample ID(s).}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
Get Image
}
\examples{
get_image()
get_image(tissueSampleId = "GTEX-1117F-0226", page = 0, itemsPerPage = 100000)
}
\concept{Histology Endpoints}
