% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{get_individuals_subset}
\alias{get_individuals_subset}
\title{Outputs sample ids for individuals with high or low ratios variability}
\usage{
get_individuals_subset(fullName, option)
}
\arguments{
\item{fullName}{full tissue name}

\item{option}{"high" or "low" - output sampl names with high or low ratios variability}
}
\value{
vector with sample ids
}
\description{
Outputs sample ids for individuals with high or low ratios variability
}
