% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linkage_disequilibrium_by_variant_data.R
\name{get_linkage_disequilibrium_by_variant_data}
\alias{get_linkage_disequilibrium_by_variant_data}
\title{Get Linkage Disequilibrium By Variant Data}
\usage{
get_linkage_disequilibrium_by_variant_data(
  variantId = NULL,
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{variantId}{String. A gtex variant ID.}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A Tibble
}
\description{
\itemize{
\item Find linkage disequilibrium (LD) data for a given variant
}

\href{https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_linkage_disequilibrium_by_variant_data_api_v2_dataset_ldByVariant_get}{GTEx Portal API documentation}
}
\examples{
get_linkage_disequilibrium_by_variant_data()
}
\seealso{
Other Datasets Endpoints: 
\code{\link{get_annotation}()},
\code{\link{get_collapsed_gene_model_exon}()},
\code{\link{get_downloads_page_data}()},
\code{\link{get_file_list}()},
\code{\link{get_full_get_collapsed_gene_model_exon}()},
\code{\link{get_functional_annotation}()},
\code{\link{get_linkage_disequilibrium_data}()},
\code{\link{get_sample}()},
\code{\link{get_subject}()},
\code{\link{get_tissue_site_detail}()},
\code{\link{get_variant_by_location}()},
\code{\link{get_variant}()}
}
\concept{Datasets Endpoints}
