% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maintenance_message.R
\name{get_maintenance_message}
\alias{get_maintenance_message}
\title{Get Maintenance Message}
\usage{
get_maintenance_message()
}
\value{
A tibble.
}
\description{
Getting all the maintenance messages from the database that are enabled.
}
\examples{
get_maintenance_message()
}
\seealso{
Other Admin Endpoints: 
\code{\link{get_news_item}()}
}
\concept{Admin Endpoints}
