% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_tissue_eqtls.R
\name{get_multi_tissue_eqtls}
\alias{get_multi_tissue_eqtls}
\title{Get Significant Single Tissue Eqtls By Location}
\usage{
get_multi_tissue_eqtls(
  gencodeId,
  variantId = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{gencodeId}{String. A Versioned GENCODE ID of a gene, e.g.
"ENSG00000065613.9".}

\item{variantId}{String. A gtex variant ID.}

\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".
Default = "gtex_v8".}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
Find significant single tissue eQTLs using Chromosomal Locations.
This service returns precomputed significant single tissue eQTLs.
Results may be filtered by tissue, and/or dataset.
By default, the service queries the latest GTEx release. Since this endpoint is used to support a third party program on the portal, the return structure is different from other endpoints and is not paginated.

\href{https://gtexportal.org/api/v2/redoc#tag/Static-Association-Endpoints/operation/get_significant_single_tissue_eqtls_api_v2_association_singleTissueEqtl_get}{GTEx Portal API documentation}
}
\examples{
\dontrun{
# search by gene
get_multi_tissue_eqtls(gencodeIds = c("ENSG00000132693.12",
                               "ENSG00000203782.5"))}
}
\seealso{
Other Static Association Endpoints: 
\code{\link{get_eqtl_genes}()},
\code{\link{get_fine_mapping}()},
\code{\link{get_independent_eqtl}()},
\code{\link{get_significant_single_tissue_eqtls}()},
\code{\link{get_sqtl_genes}()}
}
\concept{Static Association Endpoints}
