% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_news_item.R
\name{get_news_item}
\alias{get_news_item}
\title{Get News Item}
\usage{
get_news_item(page = 0, itemsPerPage = 250)
}
\arguments{
\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
Getting all the news items from the database that are current.
}
\examples{
get_news_item()
}
