% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{get_ratio_sd}
\alias{get_ratio_sd}
\title{Get allelic ratio's standard deviation for a tissue, individual-wise}
\usage{
get_ratio_sd(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
dataframe
}
\description{
Get allelic ratio's standard deviation for a tissue, individual-wise
}
