% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{get_ratios_sample}
\alias{get_ratios_sample}
\title{Gets ratios for a given tissue}
\usage{
get_ratios_sample(fullName, n_samples = 10000)
}
\arguments{
\item{fullName}{full tissue name}

\item{n_samples}{number of rows to get}
}
\value{
dataframe
}
\description{
Gets ratios for a given tissue
}
