% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{get_rin_correlations_for_tissues}
\alias{get_rin_correlations_for_tissues}
\title{Produce plot with RIN vs SD(allelic_ratio) for several tissues}
\usage{
get_rin_correlations_for_tissues()
}
\value{
ggplot2
}
\description{
Produce plot with RIN vs SD(allelic_ratio) for several tissues
}
