% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample.R
\name{get_sample_datasets_endpoints}
\alias{get_sample_datasets_endpoints}
\title{Get Sample (Datasets Endpoints)}
\usage{
get_sample_datasets_endpoints(
  datasetId = "gtex_v8",
  sampleIds = NULL,
  tissueSampleIds = NULL,
  subjectIds = NULL,
  ageBrackets = NULL,
  sex = NULL,
  pathCategory = NULL,
  tissueSiteDetailId = NULL,
  aliquotIds = NULL,
  autolysisScores = NULL,
  hardyScales = NULL,
  ischemicTime = NULL,
  ischemicTimeGroups = NULL,
  rin = NULL,
  uberonIds = NULL,
  dataTypes = NULL,
  sortBy = NULL,
  sortDirection = NULL,
  page = NULL,
  itemsPerPage = NULL
)
}
\arguments{
\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".}

\item{sampleIds}{Character vector. GTEx sample ID.}

\item{tissueSampleIds}{Array of strings. A list of Tissue Sample ID(s).}

\item{subjectIds}{Character vector. GTEx subject ID.}

\item{ageBrackets}{The age bracket(s) of the donors of interest. Options:
"20-29", "30-39", "40-49", "50-59", "60-69", "70-79".}

\item{sex}{String. Options: "male", "female".}

\item{pathCategory}{Character vector. Options: "adenoma", "amylacea",
"atelectasis", "atherosclerosis", "atherosis", "atrophy", "calcification",
"cirrhosis", "clean_specimens", "congestion", "corpora_albicantia", "cyst",
"desquamation", "diabetic", "dysplasia", "edema", "emphysema",
"esophagitis", "fibrosis", "gastritis", "glomerulosclerosis", "goiter",
"gynecomastoid", "hashimoto", "heart_failure_cells", "hemorrhage",
"hepatitis", "hyalinization", "hypereosinophilia", "hyperplasia",
"hypertrophy", "hypoxic", "infarction", "inflammation", "ischemic_changes",
"macrophages", "mastopathy", "metaplasia", "monckeberg", "necrosis",
"nephritis", "nephrosclerosis", "no_abnormalities", "nodularity",
"pancreatitis", "pigment", "pneumonia", "post_menopausal", "prostatitis",
"saponification", "scarring", "sclerotic", "solar_elastosis",
"spermatogenesis", "steatosis", "sweat_glands", "tma".}

\item{tissueSiteDetailId}{String. The ID of the tissue of interest. Can be a
GTEx specific ID (e.g. "Whole_Blood"; use \code{\link[=available_tissueSiteDetailIds]{available_tissueSiteDetailIds()}}
to see valid values) or an Ontology ID.}

\item{aliquotIds}{Character vector.}

\item{autolysisScores}{Character vector. Options: "None", "Mild", "Moderate",
"Severe".}

\item{hardyScales}{Character vector. A list of Hardy Scale(s) of interest.
Options: "Ventilator case", "Fast death - violent", "Fast death - natural
causes", "Intermediate death", "Slow death".}

\item{ischemicTime}{Integer.}

\item{ischemicTimeGroups}{Character vector. Options:
"<= 0", "1 - 300", "301 - 600", "601 - 900", "901 - 1200", "1201 - 1500", "> 1500".}

\item{rin}{Integer vector.}

\item{uberonIds}{Character vector. A list of Uberon ID(s) of interest.
Options: "UBERON:EFO_0000572", "UBERON:EFO_0002009", "UBERON:0000007",
"UBERON:0000458", "UBERON:0000473", "UBERON:0000945", "UBERON:0000992",
"UBERON:0000995", "UBERON:0000996", "UBERON:0001114", "UBERON:0001150",
"UBERON:0001157", "UBERON:0001159", "UBERON:0001211", "UBERON:0001225",
"UBERON:0001255", "UBERON:0001293", "UBERON:0001323", "UBERON:0001496",
"UBERON:0001621", "UBERON:0001870", "UBERON:0001873", "UBERON:0001874",
"UBERON:0001876", "UBERON:0001882", "UBERON:0001898", "UBERON:0001954",
"UBERON:0002037", "UBERON:0002038", "UBERON:0002046", "UBERON:0002106",
"UBERON:0002190", "UBERON:0002367", "UBERON:0002369", "UBERON:0003889",
"UBERON:0004264", "UBERON:0004550", "UBERON:0004648", "UBERON:0006330",
"UBERON:0006469", "UBERON:0006566", "UBERON:0006631", "UBERON:0006920",
"UBERON:0007610", "UBERON:0008367", "UBERON:0008952", "UBERON:0009834",
"UBERON:0009835", "UBERON:0010414", "UBERON:0011907", "UBERON:0012249",
"UBERON:0013756", "UBERON:0036149".}

\item{dataTypes}{Character vector. Options: "RNASEQ", "WGS", "WES", "OMNI",
"EXCLUDE".}

\item{sortBy}{String. Options: "sampleId", "ischemicTime", "aliquotId",
"tissueSampleId", "hardyScale", "pathologyNotes", "ageBracket",
"tissueSiteDetailId", "sex".}

\item{sortDirection}{String. Options: "asc", "desc". Default = "asc".}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble
}
\description{
This service returns information of samples used in analyses
from all datasets. Results may be filtered by dataset ID, sample ID, subject
ID, sample metadata, or other provided parameters. By default, this service
queries the latest GTEx release.

\href{https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_sample_api_v2_dataset_sample_get}{GTEx Portal API documentation}
}
\examples{
\dontrun{
 get_sample_datasets_endpoints()
}
}
\seealso{
Other Datasets Endpoints: 
\code{\link{get_annotation}()},
\code{\link{get_collapsed_gene_model_exon}()},
\code{\link{get_downloads_page_data}()},
\code{\link{get_file_list}()},
\code{\link{get_full_get_collapsed_gene_model_exon}()},
\code{\link{get_functional_annotation}()},
\code{\link{get_linkage_disequilibrium_by_variant_data}()},
\code{\link{get_linkage_disequilibrium_data}()},
\code{\link{get_subject}()},
\code{\link{get_tissue_site_detail}()},
\code{\link{get_variant_by_location}()},
\code{\link{get_variant}()}
}
\concept{Datasets Endpoints}
