% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{get_samples_tissues}
\alias{get_samples_tissues}
\title{Reads ratios from the file, subsamples and makes a plot}
\usage{
get_samples_tissues()
}
\value{
None
}
\description{
Reads ratios from the file, subsamples and makes a plot
}
