% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_NPCs.R
\name{get_sd_ratios}
\alias{get_sd_ratios}
\title{Calcultates SDs for allelic ratios}
\usage{
get_sd_ratios(df, y)
}
\arguments{
\item{df}{dataframe with counts and ratios,}

\item{y}{number of combinations}
}
\value{
ggplot
}
\description{
Calcultates SDs for allelic ratios
}
