% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_service_info.R
\name{get_service_info}
\alias{get_service_info}
\title{Get Service Info}
\usage{
get_service_info()
}
\value{
A tibble.
}
\description{
General information about the GTEx service.
}
\examples{
get_service_info()
}
