% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_short_name}
\alias{get_short_name}
\title{Gets short tissue name from a full name}
\usage{
get_short_name(fullName)
}
\arguments{
\item{fullName}{Full tissue name}
}
\value{
short tissue name
}
\description{
Gets short tissue name from a full name
}
