% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant_single_tissue_eqtls.R
\name{get_significant_single_tissue_eqtls}
\alias{get_significant_single_tissue_eqtls}
\title{Get Significant Single Tissue Eqtls}
\usage{
get_significant_single_tissue_eqtls(
  gencodeId = NULL,
  variantId = NULL,
  tissueSiteDetailId = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{gencodeId}{String, vector.}

\item{variantId}{String.}

\item{tissueSiteDetailId}{String.}

\item{datasetId}{String.}

\item{page}{Integer.}

\item{itemsPerPage}{Integer.}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Static-Association-Endpoints/operation/get_significant_single_tissue_eqtls_api_v2_association_singleTissueEqtl_get}{Find significant single tissue eQTLs.}
}
\details{
\itemize{
\item This service returns precomputed significant single tissue eQTLs.
\item Results may be filtered by tissue, gene, variant or dataset.
\item To search by gene, use the versioned GENCODE ID.
\item To search by variant, use the gtex variant ID.
}

\strong{Note:} although the GTEx Portal API documentation says to use the dbSNP rsID
when searching by variant, this returns no results. Instead use gtex variant
IDs e.g. use "chr1_153209640_C_A_b38" instead of "rs1410858".
}
\examples{
# search by gene
get_significant_single_tissue_eqtls(gencodeId = c("ENSG00000132693.12",
                                                  "ENSG00000203782.5"))

# search by variant - must be variantId (not rsid)
get_significant_single_tissue_eqtls(variantId = "chr1_153209640_C_A_b38")

# filter by gene/variant and tissue site - either `gencodeId` or `variantId`
# should be supplied as a minimum
get_significant_single_tissue_eqtls(gencodeId = c("ENSG00000132693.12",
                                                  "ENSG00000203782.5"),
                                    variantId = "chr1_153209640_C_A_b38",
                                    tissueSiteDetailId = "Whole_Blood")
}
