% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant_single_tissue_sqtls.R
\name{get_significant_single_tissue_sqtls}
\alias{get_significant_single_tissue_sqtls}
\title{Get Significant Single Tissue Sqtls}
\usage{
get_significant_single_tissue_sqtls(
  gencodeIds,
  variantId = NULL,
  tissueSiteDetailId = NULL,
  datasetId = "gtex_v8"
)
}
\arguments{
\item{gencodeIds}{A character vector of Versioned GENCODE IDs, e.g.
c("ENSG00000132693.12", "ENSG00000203782.5").}

\item{variantId}{String. A gtex variant ID.}

\item{tissueSiteDetailId}{String. The ID of the tissue of interest. Can be a
GTEx specific ID (e.g. "Whole_Blood"; use \code{\link[=available_tissueSiteDetailIds]{available_tissueSiteDetailIds()}}
to see valid values) or an Ontology ID.}

\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".}
}
\value{
A tibble
}
\description{
Retrieve Single Tissue sQTL Data. This service returns single tissue sQTL
data for the given genes, from a specified dataset. Results may be filtered
by tissue By default, the service queries the latest GTEx release.
}
\details{
The retrieved data is split into pages with items_per_page entries per page

\href{https://gtexportal.org/api/v2/redoc#tag/Static-Association-Endpoints/operation/get_significant_single_tissue_sqtls_api_v2_association_singleTissueSqtl_get}{GTEx Portal API documentation}
}
\examples{
# search by gene
get_significant_single_tissue_sqtls(gencodeIds = c("ENSG00000065613.9",
                                                   "ENSG00000203782.5"))
}
\seealso{
Other Static Association Endpoints: 
\code{\link{get_eqtl_genes}()},
\code{\link{get_fine_mapping}()},
\code{\link{get_independent_eqtl}()},
\code{\link{get_multi_tissue_eqtls}()},
\code{\link{get_significant_single_tissue_eqtls_by_location}()},
\code{\link{get_significant_single_tissue_eqtls}()},
\code{\link{get_significant_single_tissue_ieqtls}()},
\code{\link{get_significant_single_tissue_isqtls}()},
\code{\link{get_sqtl_genes}()}
}
\concept{Static Association Endpoints}
