% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tissue_code}
\alias{get_tissue_code}
\title{Gets tissue code from a short name}
\usage{
get_tissue_code(name)
}
\arguments{
\item{name}{Short tissue name}
}
\value{
tissue code
}
\description{
Gets tissue code from a short name
}
