% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant.R
\name{get_variant}
\alias{get_variant}
\title{Get Variant}
\usage{
get_variant(
  snpId = NULL,
  variantId = NULL,
  chromosome = NULL,
  pos = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{snpId}{String}

\item{variantId}{String. A gtex variant ID.}

\item{chromosome}{String. One of "chr1", "chr2", "chr3", "chr4", "chr5",
"chr6", "chr7", "chr8", "chr9", "chr10", "chr11", "chr12", "chr13",
"chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21",
"chr22", "chrM", "chrX", "chrY".}

\item{pos}{Integer, vector.}

\item{datasetId}{String (default = "gtex_v8"). Options: "gtex_v8" or
"gtex_snrnaseq_pilot". Unique identifier of a dataset. Usually includes a
data source and data release.}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_variant_api_v2_dataset_variant_get}{This service returns information about a variant, including position, dbSNP RS ID, the reference allele, the alternative allele, and whether the minor allele frequency is >= 1\%.}
}
\details{
For GTEx v6p, there is also information about whether the whole exome
sequence and chip sequencing data are available. Results may be queried by
GTEx variant ID (variantId), dbSNP RS ID (snpId) or genomic location
(chromosome and pos). Variants are identified based on the genotype data of
each dataset cohort, namely, are dataset-dependent. Each variant is assigned
a unique GTEx variant ID (i.e. the primary key). Not all variants have a
mappable dbSNP RS ID. By default, this service queries the latest GTEx
release.

Note: searching by chromosome and position can be more easily achieved using
\code{\link[=get_variant_by_location]{get_variant_by_location()}}.
}
\examples{
# search by rsid
get_variant(snpId = "rs1410858")

# search by variantId
get_variant(variantId = "chr1_153209640_C_A_b38")

# search by chromosome and position
get_variant(chromosome = "chr1",
            pos = 153209600:153209700)
}
\seealso{
Other Datasets Endpoints: 
\code{\link{get_variant_by_location}()}
}
\concept{Datasets Endpoints}
