% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtexr-package.R
\docType{package}
\name{gtexr-package}
\alias{gtexr}
\alias{gtexr-package}
\title{gtexr: Query the GTEx Portal API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A convenient R interface to the GTEx Portal API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/gtexr/}
  \item \url{https://github.com/rmgpanw/gtexr}
  \item Report bugs at \url{https://github.com/rmgpanw/gtexr/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID}) [copyright holder]

}
\keyword{internal}
