% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{list_of_samples_with_ratios}
\alias{list_of_samples_with_ratios}
\title{Run ratios analysis for NPCs}
\usage{
list_of_samples_with_ratios(sample_id)
}
\arguments{
\item{sample_id}{sample id}
}
\value{
dataframe
}
\description{
Run ratios analysis for NPCs
}
