% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_counts.R
\name{load_counts}
\alias{load_counts}
\title{Loads counts from Rda}
\usage{
load_counts(fullName,
  path_rda = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/data_Rda/")
}
\arguments{
\item{fullName}{tissue name}

\item{path_rda}{Path to the folder contatining all .Rda files for tissues}
}
\value{
ExpressionSet
}
\description{
Loads counts from Rda
}
