% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGtexData.R
\name{load_Gtex_data}
\alias{load_Gtex_data}
\title{Coverts GTEx data from 3 files (downloaded from GTEx website) to ExpressionSet format}
\usage{

  load_Gtex_data(path_to_files = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/raw_data")
}
\arguments{
\item{path_to_files}{Path to the input folder contatining all 3 files}
}
\value{
ExpressionSet
}
\description{
This function loads files with GTEx subjects phenotypes, attributes and RNA-seq counts
and coverts them to one ExpressionSet objects.
}
