% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_NPCs.R
\name{measure}
\alias{measure}
\title{Calculates CVs for "observed" BAE and all genes from counts and ratios}
\usage{
measure(a)
}
\arguments{
\item{a}{vector, first 8 entries are counts, second 8 entries are ratios}
}
\value{
vector with CVs and mean
}
\description{
Calculates CVs for "observed" BAE and all genes from counts and ratios
}
