% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCV2_vs_mean.R
\name{plot_CV2_vs_mean}
\alias{plot_CV2_vs_mean}
\title{Plots CV2 versus mean}
\usage{
plot_CV2_vs_mean(df, name,
  path_fig = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/results/Figures/CV2_mean/",
  n_bins = 5, plotMode = "Both", modeBoxPlot = "None")
}
\arguments{
\item{df}{dataframe with statistics}

\item{name}{tissue name}

\item{path_fig}{path to save figures}

\item{n_bins}{number of bins for mean expression}

\item{plotMode}{"Both" or "Boxplots" or "Scatterplot" or "Merged" (merge boxplot and sctterplot into one plot)}

\item{modeBoxPlot}{"None", "Ranges" (bin ranges will be added to boxplots), "Sizes" (sample sizes will be added to the plot) or "Pvals" (pvalues will be added to boxplots)}
}
\value{
ggplot
}
\description{
Plots CV2 versus mean
}
