% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{plot_density_ratio_per_individual}
\alias{plot_density_ratio_per_individual}
\alias{plot_density_ratio_per_individual}
\title{Plot ratio densities for all tissues}
\usage{
plot_density_ratio_per_individual(fullName)

plot_density_ratio_per_individual(fullName)
}
\arguments{
\item{fullName}{full name of the tissue}
}
\value{
ggplot2

ggplot2
}
\description{
Plot ratio densities for all tissues

Reads ratios from the file and makes a plot of density, colored by individual
}
