% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{plot_density_ratio_per_individual_all_tissues}
\alias{plot_density_ratio_per_individual_all_tissues}
\title{Plot ratio densities for all tissues}
\usage{
plot_density_ratio_per_individual_all_tissues()
}
\value{
ggplot2
}
\description{
Plot ratio densities for all tissues
}
