% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCV2_vs_mean.R
\name{plot_one}
\alias{plot_one}
\title{Plots boxplots for ranges}
\usage{
plot_one(df, n_bins, qq, modeBoxPlot = "None")
}
\arguments{
\item{df}{dataframe with statistics}

\item{n_bins}{number of bins for mean expression}

\item{qq}{quantiles}

\item{modeBoxPlot}{"None", "Ranges" (bin ranges will be added to boxplots), "Sizes" (sample sizes will be added to the plot) or "Pvals" (pvalues will be added to boxplots)}
}
\value{
ggplot
}
\description{
Plots boxplots for ranges
}
