% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCV2_vs_mean.R
\name{plot_two}
\alias{plot_two}
\title{Plots scatterplots}
\usage{
plot_two(df, n_bins, qq)
}
\arguments{
\item{df}{dataframe with statistics}

\item{n_bins}{number of bins for mean expression}

\item{qq}{quantiles}
}
\value{
ggplot
}
\description{
Plots scatterplots
}
