% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_counts.R
\name{process_to_counts}
\alias{process_to_counts}
\title{Proccesses GTEx counts and merges counts statistics with classification data}
\usage{
process_to_counts(name, keepCounts = FALSE, sample_names = "all",
  mode = "norm_counts",
  path_to_class = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/classification/",
  path_rda = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/data_Rda/")
}
\arguments{
\item{name}{Tissue name}

\item{keepCounts}{Keep columns with counts in the output}

\item{sample_names}{Default is set to all meaning "all samples"; can be a string of names (c("GTEX-111CU-1326-SM-5NQ8L", "GTEX-111YS-1426-SM-5GID8", "GTEX-1122O-1326-SM-5H11F"))}

\item{mode}{"norm_counts" / "log" meaning using normalized by library size counts or using log(normalized_counts)}

\item{path_to_class}{Path to file name_class.txt with MAE/BAE classification call (output by MAGIC or RNA-seq analysis)}

\item{path_rda}{Path to the folder contatining all .Rda files for tissues}
}
\value{
dataframe with MAE/BAE classification and counts statistics
}
\description{
Proccesses GTEx counts and merges counts statistics with classification data
}
