% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{read_counts}
\alias{read_counts}
\title{Reads counts for a given tissue}
\usage{
read_counts(fullName, coverage = 0,
  path = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/processed_data/tissue_AI_counts_upd/")
}
\arguments{
\item{fullName}{full tissue name}

\item{coverage}{Minimum number of counts per gene}

\item{path}{path with the tissues' ratios}
}
\value{
None
}
\description{
Reads counts for a given tissue
}
