% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{read_counts_and_calc_ratios}
\alias{read_counts_and_calc_ratios}
\title{Reads counts for a given tissue, aggregates them and puts together ratios for all samples}
\usage{
read_counts_and_calc_ratios(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
dataframe
}
\description{
Reads counts for a given tissue, aggregates them and puts together ratios for all samples
}
