% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{read_ratios}
\alias{read_ratios}
\title{Reads ratios for a given tissue}
\usage{
read_ratios(fullName, coverage = 0,
  path = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/processed_data/tissue_AI_ratios_upd/")
}
\arguments{
\item{fullName}{full tissue name}

\item{coverage}{coverage threshold}

\item{path}{path with the tissues' ratios}
}
\value{
None
}
\description{
Reads ratios for a given tissue
}
