% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{save_counts}
\alias{save_counts}
\title{Reads and saves counts for a given tissue}
\usage{
save_counts(fullName,
  path = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/processed_data/tissue_AI_counts_upd/")
}
\arguments{
\item{fullName}{full tissue name}

\item{path}{path to save the tissues' ratios}
}
\value{
None
}
\description{
Reads and saves counts for a given tissue
}
