% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{save_counts_tissues_of_interest}
\alias{save_counts_tissues_of_interest}
\title{Saves counts to files for the tissues of interest}
\usage{
save_counts_tissues_of_interest()
}
\description{
Saves counts to files for the tissues of interest
}
