% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGtexData.R
\name{save_Gtex_data_by_tissue}
\alias{save_Gtex_data_by_tissue}
\title{Saves selected tissues counts to .Rda files}
\usage{

  save_Gtex_data_by_tissue(path_to_raw_files = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/raw_data",
  path_outFolder = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/data_Rda_upd/",
  samples_thr = 30)
}
\arguments{
\item{path_to_raw_files}{Path to raw GTEx files}

\item{path_outFolder}{Path to folder to save Rda files for tissues}

\item{number_threshold}{Minimum number of samples per tissue to consider the tissue}
}
\description{
Saves selected tissues counts to .Rda files
}
