% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{save_ratios}
\alias{save_ratios}
\title{Reads and saves ratios for a given tissue}
\usage{
save_ratios(fullName,
  path = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/processed_data/tissue_AI_ratios/")
}
\arguments{
\item{fullName}{full tissue name}

\item{path}{path to save the tissues' ratios}
}
\description{
Reads and saves ratios for a given tissue
}
