% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAI_Data.R
\name{save_ratios_tissues_of_interest}
\alias{save_ratios_tissues_of_interest}
\title{Saves ratios to files for the tissues of interest}
\usage{
save_ratios_tissues_of_interest()
}
\value{
None
}
\description{
Saves ratios to files for the tissues of interest
}
