% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_NPCs.R
\name{simulate_ratios}
\alias{simulate_ratios}
\title{Makes boxplots for SDs}
\usage{
simulate_ratios(df)
}
\arguments{
\item{df}{dataframe with counts and ratios}
}
\value{
ggplot
}
\description{
Makes boxplots for SDs
}
