% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_NPCs.R
\name{subset_and_get_ratio}
\alias{subset_and_get_ratio}
\title{Pulls data and gets allelic ratios}
\usage{
subset_and_get_ratio(df, x)
}
\arguments{
\item{df}{dataframe with counts and ratios,}

\item{x}{vector with columns to pull}
}
\value{
ggplot
}
\description{
Pulls data and gets allelic ratios
}
