% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGtexData.R
\name{subsetGtexData}
\alias{subsetGtexData}
\title{Subsets GTEx ExpressionSet to a given tissue}
\usage{
subsetGtexData(gtex_es, name)
}
\arguments{
\item{gtex_es}{ExpressionSet with GTEx data}

\item{name}{Full tissue name to subset to}
}
\value{
ExpressionSet
}
\description{
Subsets GTEx ExpressionSet to a given tissue
}
