% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{variation_analysis_for_subset}
\alias{variation_analysis_for_subset}
\title{Do variation analysis for samples with moderate sd values}
\usage{
variation_analysis_for_subset(fullName)
}
\arguments{
\item{fullName}{full tissue name}
}
\value{
None
}
\description{
Do variation analysis for samples with moderate sd values
}
