% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runForTissue.R
\name{zscores_heatmap}
\alias{zscores_heatmap}
\title{Convert normalized counts to z-scores and plot heatmaps}
\usage{
zscores_heatmap(name, sortedBy = "CV2")
}
\arguments{
\item{name}{Tissue name}

\item{sortedBy}{CV2 or status}
}
\value{
None
}
\description{
Convert normalized counts to z-scores and plot heatmaps
}
