% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation.R
\name{get_annotation}
\alias{get_annotation}
\title{Get Annotation}
\usage{
get_annotation(datasetId = "gtex_v8", page = 0, itemsPerPage = 250)
}
\arguments{
\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble
}
\description{
This service returns the list of annotations and allowed values
by which a particular dataset can be subsetted. Results may be filtered by
dataset.

\href{https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_annotation_api_v2_dataset_annotation_get}{GTEx Portal API documentation}
}
\details{
Note: the output for this function appears to be incomplete
currently.
}
\examples{
\dontrun{
 get_annotation()
}
}
\seealso{
Other Datasets Endpoints: 
\code{\link{get_collapsed_gene_model_exon}()},
\code{\link{get_downloads_page_data}()},
\code{\link{get_file_list}()},
\code{\link{get_full_get_collapsed_gene_model_exon}()},
\code{\link{get_functional_annotation}()},
\code{\link{get_linkage_disequilibrium_by_variant_data}()},
\code{\link{get_linkage_disequilibrium_data}()},
\code{\link{get_sample_datasets_endpoints}()},
\code{\link{get_subject}()},
\code{\link{get_tissue_site_detail}()},
\code{\link{get_variant}()},
\code{\link{get_variant_by_location}()}
}
\concept{Datasets Endpoints}
