% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_counts.R
\name{get_counts_statistics}
\alias{get_counts_statistics}
\title{Calculates mean, sd, CV, CV2 for each gene}
\usage{
get_counts_statistics(sv, min_counts_per_gene = 10, keepCounts = FALSE,
  mode = "norm_counts")
}
\arguments{
\item{sv}{ExpressionSet for a given tissue}

\item{min_counts_per_gene}{Minimum number of counts per gene to be considered}

\item{keepCounts}{Keep columns with counts in the output}

\item{mode}{"norm_counts" / "log" meaning using normalized by library size counts or using log(normalized_counts)}
}
\value{
dataframe with counts statistics
}
\description{
Calculates mean, sd, CV, CV2 for each gene
}
