% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get Dataset Info}
\usage{
get_dataset_info(datasetId = "gtex_v8", organizationName = "GTEx Consortium")
}
\arguments{
\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".}

\item{organizationName}{String. Options: "GTEx Consortium" "Kid's First".}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Metadata-Endpoints/operation/get_dataset_info_api_v2_metadata_dataset_get}{GTEx Portal API documentation}
}
\details{
Note: this current only returns empty values.
}
\examples{
\dontrun{
get_dataset_info()
}
}
\concept{Metadata Endpoints}
