% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get Dataset Info}
\usage{
get_dataset_info(
  datasetId = NULL,
  organizationName = NULL,
  .return_raw = FALSE
)
}
\arguments{
\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".}

\item{organizationName}{String. Options: "GTEx Consortium" "Kid's First".}

\item{.return_raw}{Logical. If \code{TRUE}, return the raw API JSON response.
Default = \code{FALSE}}
}
\value{
A tibble. Or a list if \code{.return_raw = TRUE}.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Metadata-Endpoints/operation/get_dataset_info_api_v2_metadata_dataset_get}{GTEx Portal API documentation}
}
\examples{
\dontrun{
get_dataset_info(datasetId = "gtex_v8", organizationName = "GTEx Consortium")
}
}
\concept{Metadata Endpoints}
