% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eqtl_genes.R
\name{get_eqtl_genes}
\alias{get_eqtl_genes}
\title{Get Eqtl Genes}
\usage{
get_eqtl_genes(
  tissueSiteDetailId,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{tissueSiteDetailId}{String. The ID of the tissue of interest. Can be a
GTEx specific ID (e.g. "Whole_Blood"; use \code{\link[=available_tissueSiteDetailIds]{available_tissueSiteDetailIds()}}
to see valid values) or an Ontology ID.}

\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".
Default = "gtex_v8".}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
Retrieve eGenes (eQTL Genes).
}
\details{
\itemize{
\item This service returns eGenes (eQTL Genes) from the specified dataset.
\item eGenes are genes that have at least one significant cis-eQTL acting upon them.
\item Results may be filtered by tissue. By default, the service queries the latest GTEx release.
}

For each eGene, the results include the allelic fold change
(log2AllelicFoldChange), p-value (pValue), p-value threshold
(pValueThreshold), empirical p-value (empiricalPValue), and q-value (qValue).
\itemize{
\item The log2AllelicFoldChange is the allelic fold change (in log2 scale) of the most significant eQTL.
\item The pValue is the nominal p-value of the most significant eQTL.
\item The pValueThreshold is the p-value threshold used to determine whether a cis-eQTL for this gene is significant. For more details see https://gtexportal.org/home/documentationPage#staticTextAnalysisMethods.
\item The empiricalPValue is the beta distribution-adjusted empirical p-value from FastQTL.
\item The qValues were calculated based on the empirical p-values. A false discovery rate (FDR) threshold of <= 0.05 was applied to identify genes with a significant eQTL.
}
}
\examples{
get_eqtl_genes("Whole_Blood")
}
\seealso{
Other Static Association Endpoints: 
\code{\link{get_fine_mapping}()},
\code{\link{get_independent_eqtl}()},
\code{\link{get_multi_tissue_eqtls}()},
\code{\link{get_significant_single_tissue_Ieqtls}()},
\code{\link{get_significant_single_tissue_eqtls_by_location}()},
\code{\link{get_significant_single_tissue_eqtls}()},
\code{\link{get_significant_single_tissue_isqtls}()},
\code{\link{get_significant_single_tissue_sqtls}()},
\code{\link{get_sqtl_genes}()}
}
\concept{Static Association Endpoints}
