% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_get_collapsed_gene_model_exon.R
\name{get_full_get_collapsed_gene_model_exon}
\alias{get_full_get_collapsed_gene_model_exon}
\title{Get Full Get Collapsed Gene Model Exon}
\usage{
get_full_get_collapsed_gene_model_exon(gencodeId, page = 0, itemsPerPage = 250)
}
\arguments{
\item{gencodeId}{String. A Versioned GENCODE ID of a gene, e.g.
"ENSG00000065613.9".}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble
}
\description{
This service allows the user to query the full Collapsed Gene
Model Exon of a specific gene by gencode ID

\href{https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_full_get_collapsed_gene_model_exon_api_v2_dataset_fullCollapsedGeneModelExon_get}{GTEx Portal API documentation}
}
\examples{
\dontrun{
  get_full_get_collapsed_gene_model_exon(gencodeId = "ENSG00000203782.5")
}
}
\seealso{
Other Datasets Endpoints: 
\code{\link{get_annotation}()},
\code{\link{get_collapsed_gene_model_exon}()},
\code{\link{get_downloads_page_data}()},
\code{\link{get_file_list}()},
\code{\link{get_functional_annotation}()},
\code{\link{get_linkage_disequilibrium_by_variant_data}()},
\code{\link{get_linkage_disequilibrium_data}()},
\code{\link{get_sample_datasets_endpoints}()},
\code{\link{get_subject}()},
\code{\link{get_tissue_site_detail}()},
\code{\link{get_variant}()},
\code{\link{get_variant_by_location}()}
}
\concept{Datasets Endpoints}
