% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Get Image}
\usage{
get_image(tissueSampleIds = NULL, page = 0, itemsPerPage = 250)
}
\arguments{
\item{tissueSampleIds}{Array of strings. A list of Tissue Sample ID(s).}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Histology-Endpoints/operation/get_image_api_v2_histology_image_get}{GTEx Portal API documentation}
}
\examples{
\dontrun{
get_image()

# filter by `tissueSampleId`
result <- get_image(tissueSampleIds = "GTEX-1117F-0526")
print(result)

# note that `pathologyNotesCategories` (if present) is a list column
print(result$pathologyNotesCategories)
}
}
\concept{Histology Endpoints}
