% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maintenance_message.R
\name{get_maintenance_message}
\alias{get_maintenance_message}
\title{Get Maintenance Message}
\usage{
get_maintenance_message(
  page = 0,
  itemsPerPage = getOption("gtexr.itemsPerPage"),
  .verbose = getOption("gtexr.verbose"),
  .return_raw = FALSE
)
}
\arguments{
\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250). Set globally to maximum value
100000 with \code{options(list(gtexr.itemsPerPage = 100000))}.}

\item{.verbose}{Logical. If \code{TRUE} (default), print paging information. Set
to \code{FALSE} globally with \code{options(list(gtexr.verbose = FALSE))}.}

\item{.return_raw}{Logical. If \code{TRUE}, return the raw API JSON response.
Default = \code{FALSE}}
}
\value{
A tibble. Or a list if \code{.return_raw = TRUE}.
}
\description{
Getting all the maintenance messages from the database that are
enabled.

\href{https://gtexportal.org/api/v2/redoc#tag/Admin-Endpoints/operation/get_maintenance_message_api_v2_admin_maintenanceMessage_get}{GTEx Portal API documentation}.
}
\details{
Note this typically returns an empty tibble.
}
\examples{
\dontrun{
get_maintenance_message()
}
}
\seealso{
Other Admin Endpoints: 
\code{\link{get_news_item}()}
}
\concept{Admin Endpoints}
