% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_news_item.R
\name{get_news_item}
\alias{get_news_item}
\title{Get News Item}
\usage{
get_news_item(page = 0, itemsPerPage = 250)
}
\arguments{
\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
Getting all the news items from the database that are current.

\href{https://gtexportal.org/api/v2/redoc#tag/Admin-Endpoints/operation/get_news_item_api_v2_admin_newsItem_get}{GTEx Portal API documentation}.
}
\examples{
\dontrun{
get_news_item()
}
}
\seealso{
Other Admin Endpoints: 
\code{\link{get_maintenance_message}()}
}
\concept{Admin Endpoints}
