% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_service_info.R
\name{get_service_info}
\alias{get_service_info}
\title{Get Service Info}
\usage{
get_service_info(.return_raw = FALSE)
}
\arguments{
\item{.return_raw}{Logical. If \code{TRUE}, return the raw API JSON response.
Default = \code{FALSE}}
}
\value{
A tibble. Or a list if \code{.return_raw = TRUE}.
}
\description{
General information about the GTEx service.

\href{https://gtexportal.org/api/v2/redoc#tag/GTEx-Portal-API-Info/operation/get_service_info_api_v2__get}{GTEx Portal API documentation}.
}
\examples{
\dontrun{
get_service_info()
}
}
\concept{GTEx Portal API Info}
