% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant.R
\name{get_variant}
\alias{get_variant}
\title{Get Variant}
\usage{
get_variant(
  snpId = NULL,
  variantId = NULL,
  chromosome = NULL,
  pos = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{snpId}{String.}

\item{variantId}{String.}

\item{chromosome}{String.}

\item{pos}{Integer, vector.}

\item{datasetId}{String.}

\item{page}{Integer.}

\item{itemsPerPage}{Integer.}
}
\value{
A tibble.
}
\description{
This service returns information about a variant, including position, dbSNP
RS ID, the reference allele, the alternative allele, and whether the minor
allele frequency is >= 1\%. For GTEx v6p, there is also information about
whether the whole exome sequence and chip sequencing data are available.
Results may be queried by GTEx variant ID (variantId), dbSNP RS ID (snpId) or
genomic location (chromosome and pos). Variants are identified based on the
genotype data of each dataset cohort, namely, are dataset-dependent. Each
variant is assigned a unique GTEx variant ID (i.e. the primary key). Not all
variants have a mappable dbSNP RS ID. By default, this service queries the
latest GTEx release.
}
\examples{
# search by rsid
get_variant(snpId = "rs1410858")

# search by variantId
get_variant(variantId = "chr1_153209640_C_A_b38")

# search by chromosome and position
get_variant(chromosome = "chr1",
            pos = 153209600:153209700)
}
