% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variant_by_location.R
\name{get_variant_by_location}
\alias{get_variant_by_location}
\title{Get Variant By Location}
\usage{
get_variant_by_location(
  start,
  end,
  chromosome,
  sortDirection = "asc",
  page = 0,
  itemsPerPage = 250
)
}
\arguments{
\item{start}{Integer.}

\item{end}{Integer.}

\item{chromosome}{String. One of "chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9", "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21", "chr22", "chrM", "chrX", "chrY".}

\item{sortDirection}{String. Options: "asc", "desc".
Default = "asc".}

\item{page}{Integer (default = 0).}

\item{itemsPerPage}{Integer (default = 250).}
}
\value{
A tibble.
}
\description{
\href{https://gtexportal.org/api/v2/redoc#tag/Datasets-Endpoints/operation/get_variant_by_location_api_v2_dataset_variantByLocation_get}{This service allows the user to query information about variants on a certain chromosome at a certain location.}
}
\examples{
get_variant_by_location(start = 153209600,
                        end = 153209700,
                        chromosome = "chr1")
}
\seealso{
Other Datasets Endpoints: 
\code{\link{get_variant}()}
}
\concept{Datasets Endpoints}
