% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtexr-package.R
\docType{package}
\name{gtexr-package}
\alias{gtexr}
\alias{gtexr-package}
\title{gtexr: Query the GTEx Portal API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API. For more information on the API, see \url{https://gtexportal.org/api/v2/redoc}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/gtexr/}
  \item \url{https://github.com/rmgpanw/gtexr}
  \item Report bugs at \url{https://github.com/rmgpanw/gtexr/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick.19@ucl.ac.uk} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Benjamin Zuckerman (\href{https://orcid.org/0000-0002-0077-6074}{ORCID})
  \item Abraham Olvera-Barrios (\href{https://orcid.org/0000-0002-3305-4465}{ORCID})
  \item Chuin Ying Ung (\href{https://orcid.org/0000-0001-8487-4589}{ORCID})
  \item Robert Luben (\href{https://orcid.org/0000-0002-5088-6343}{ORCID})
}

}
\keyword{internal}
