% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtexr_arguments.R
\name{gtexr_arguments}
\alias{gtexr_arguments}
\title{gtexr arguments}
\usage{
gtexr_arguments()
}
\arguments{
\item{bp_window}{Integer.}

\item{chromosome}{String. One of "chr1", "chr2", "chr3", "chr4", "chr5",
"chr6", "chr7", "chr8", "chr9", "chr10", "chr11", "chr12", "chr13",
"chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21",
"chr22", "chrM", "chrX", "chrY".}

\item{datasetId}{String. Unique identifier of a dataset. Usually includes a
data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".
Default = "gtex_v8".}

\item{end}{Integer.}

\item{gencodeId}{String. A Versioned GENCODE ID of a gene, e.g.
"ENSG00000065613.9".}

\item{gencodeIds}{A character vector of Versioned GENCODE IDs, e.g.
c("ENSG00000132693.12", "ENSG00000203782.5").}

\item{gencodeVersion}{String (default = "v26"). GENCODE annotation release.
Either "v26" or "v19".}

\item{geneId}{String. A gene symbol, a gencode ID, or an Ensemble ID.}

\item{geneIds}{A character vector of gene symbols, versioned gencodeIds, or
unversioned gencodeIds.}

\item{genomeBuild}{String. Options: "GRCh38/hg38", "GRCh37/hg19". Default =
"GRCh38/hg38".}

\item{itemsPerPage}{Integer (default = 250).}

\item{organizationName}{String. Options: "GTEx Consortium" "Kid's First".}

\item{page}{Integer (default = 0).}

\item{pos}{Integer, vector.}

\item{snpId}{String}

\item{sortDirection}{String. Options: "asc", "desc". Default = "asc".}

\item{start}{Integer.}

\item{tissueSiteDetailId}{String. The ID of the tissue of interest. Can be a
GTEx specific ID (e.g. "Whole_Blood"; use \code{\link[=available_tissueSiteDetailIds]{available_tissueSiteDetailIds()}}
to see valid values) or an Ontology ID.}

\item{tissueSiteDetailIds}{Character vector of IDs for tissues of interest.
Can be GTEx specific IDs (e.g. "Whole_Blood"; use
\code{\link[=available_tissueSiteDetailIds]{available_tissueSiteDetailIds()}} to see valid values) or Ontology IDs.}

\item{variantId}{String. A gtex variant ID.}
}
\value{
Returns \code{NULL} invisibly. Used for documentation only.
}
\description{
Internal function that documents all arguments for exported gtexr functions
that wrap GTEx Portal API endpoints (using roxygen \verb{@inheritParams} tag).
}
\keyword{internal}
