% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processToCounts.R
\name{process_to_counts}
\alias{process_to_counts}
\title{Proccesses GTEx counts and merges counts statistics with classification data}
\usage{
process_to_counts(name,
  path_to_class = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/classification/",
  path_rda = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/data_Rda_upd/")
}
\arguments{
\item{name}{Tissue name}

\item{path_to_class}{Path to file name_class.txt with MAE/BAE classification call (output by MAGIC or RNA-seq analysis)}

\item{path_rda}{Path to the folder contatining all .Rda files for tissues}
}
\value{
dataframe with MAE/BAE classification and counts statistics
}
\description{
Proccesses GTEx counts and merges counts statistics with classification data
}
