% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{read_counts}
\alias{read_counts}
\title{Reads counts for a given tissue}
\usage{
read_counts(fullName,
  path = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/processed_data/tissue_AI_counts/")
}
\arguments{
\item{fullName}{full tissue name}

\item{path}{path with the tissues' ratios}
}
\description{
Reads counts for a given tissue
}
