% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_tissue.R
\name{read_ratios}
\alias{read_ratios}
\title{Reads ratios for a given tissue}
\usage{
read_ratios(fullName,
  path = "/Users/svetlana/Dropbox (Partners HealthCare)/variation_project/GTEx/GTExR/data/processed_data/tissue_AI_ratios/")
}
\arguments{
\item{fullName}{full tissue name}

\item{path}{path with the tissues' ratios}
}
\description{
Reads ratios for a given tissue
}
